CDCAirbox_Linebot/requirements.txt
line-bot-sdk
flask
pymongo
numpy
requests
psycopg2-binary
python-dotenv
